Render values and animate them
The animations are not the way they should be yet.  Ideally we'd have
both a fade and a move action.  The move in itself is of high value
already.  We do automatically limit the amount of values to the
selection that was made.

Altering this model, we may want to treat the year in which the
valuation was made as a separate predicate and class, so we can loop
over the years in a special manner.  That being said, we would not
have been able to test with the limited dataset we currently have
available.